LayoutTests/imported/w3c/web-platform-tests/speculation-rules/prefetch/referrer-policy-from-rules.https_7-7-expected.txt


FAIL with "unsafe-url" referrer policy in rule set overriding "strict-origin" of cross-site referring page assert_equals: must not be prefetched expected (undefined) undefined but got (string) "prefetch"
